add functional components lesson_1_2
